Prevent from releasing empty version numbers
When incrementing version number, both semver 1.4.x and 2.0.10 returns a
_null_ value if something went wrong.

As far as I know, the two main cases leading to a _null_ resulting
version are either a bad initial version number, or a bad release type
(e.g. _majour_).

This patch just makes grunt-release check the resulting version number
(either after a bump or not) for non nullness and throw a grunt warning
if the test fails.

Another solution would have been to add many sanity checks before
starting the release task like checking that the provided release type
has expected value (major, minor, patch, etc) but it appears it is more
flexible to just let the _semver_ library to do the hard work for us and
simply check for non nullness of the resulting version.